+++ Adapters/IronSource/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## IronSource iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationIronSource', '9.4.0.1'
+pod 'CASMediationIronSource', '9.4.1.0'
 ```
+
+### 9.4.1.0
+- Certified with IronSource - 9.4.1
 
 ### 9.4.0.1
 - Fixed a rare issue with BURL being called for the wrong ad source.