+++ Adapters/Pangle/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Pangle iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationPangle', '8.0.0.7.0'
+pod 'CASMediationPangle', '8.0.0.8.0'
 ```
+
+### 8.0.0.8.0
+- Certified with Pangle - 8.0.0.8
 
 ### 8.0.0.7.0
 - Certified with Pangle - 8.0.0.7